chore: migrate to app router
