+++ app/dashboard/layout.tsx
@@ -1,0 +1,17 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Privy Auth Demo",
+};
+
+export default function DashboardLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      <main>{children}</main>
+    </>
+  );
+}
