+++ app/dashboard/page.tsx
@@ -1,0 +1,222 @@
+"use client";
+
+import { getAccessToken, usePrivy } from "@privy-io/react-auth";
+import { useRouter } from "next/navigation";
+import { useEffect, useState } from "react";
+import WalletList from "../../components/WalletList";
+
+export default function DashboardPage() {
+  const [verifyResult, setVerifyResult] = useState();
+  const router = useRouter();
+  const {
+    ready,
+    authenticated,
+    user,
+    logout,
+    linkEmail,
+    linkWallet,
+    unlinkEmail,
+    linkPhone,
+    unlinkPhone,
+    unlinkWallet,
+    linkGoogle,
+    unlinkGoogle,
+    linkTwitter,
+    unlinkTwitter,
+    linkDiscord,
+    unlinkDiscord,
+  } = usePrivy();
+
+  useEffect(() => {
+    if (ready && !authenticated) {
+      router.push("/");
+    }
+  }, [ready, authenticated, router]);
+
+  const numAccounts = user?.linkedAccounts?.length || 0;
+  const canRemoveAccount = numAccounts > 1;
+
+  const email = user?.email;
+  const phone = user?.phone;
+  const wallet = user?.wallet;
+
+  const googleSubject = user?.google?.subject || null;
+  const twitterSubject = user?.twitter?.subject || null;
+  const discordSubject = user?.discord?.subject || null;
+
+  async function verifyToken() {
+    const accessToken = await getAccessToken();
+    const result = await fetch("/api/verify", {
+      headers: {
+        ...(accessToken ? { Authorization: `Bearer ${accessToken}` } : {}),
+      },
+    });
+
+    return await result.json();
+  }
+
+  return (
+    <main className="flex flex-col min-h-screen px-4 sm:px-20 py-6 sm:py-10 bg-privy-light-blue">
+      {ready && authenticated ? (
+        <>
+          <div className="flex flex-row justify-between">
+            <h1 className="text-2xl font-semibold">Privy Auth Demo</h1>
+            <button
+              onClick={logout}
+              className="text-sm bg-violet-200 hover:text-violet-900 py-2 px-4 rounded-md text-violet-700"
+            >
+              Logout
+            </button>
+          </div>
+          <div className="mt-12 flex gap-4 flex-wrap">
+            {googleSubject ? (
+              <button
+                onClick={() => {
+                  unlinkGoogle(googleSubject);
+                }}
+                className="text-sm border border-violet-600 hover:border-violet-700 py-2 px-4 rounded-md text-violet-600 hover:text-violet-700 disabled:border-gray-500 disabled:text-gray-500 hover:disabled:text-gray-500"
+                disabled={!canRemoveAccount}
+              >
+                Unlink Google
+              </button>
+            ) : (
+              <button
+                onClick={() => {
+                  linkGoogle();
+                }}
+                className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white"
+              >
+                Link Google
+              </button>
+            )}
+
+            {twitterSubject ? (
+              <button
+                onClick={() => {
+                  unlinkTwitter(twitterSubject);
+                }}
+                className="text-sm border border-violet-600 hover:border-violet-700 py-2 px-4 rounded-md text-violet-600 hover:text-violet-700 disabled:border-gray-500 disabled:text-gray-500 hover:disabled:text-gray-500"
+                disabled={!canRemoveAccount}
+              >
+                Unlink Twitter
+              </button>
+            ) : (
+              <button
+                className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white"
+                onClick={() => {
+                  linkTwitter();
+                }}
+              >
+                Link Twitter
+              </button>
+            )}
+
+            {discordSubject ? (
+              <button
+                onClick={() => {
+                  unlinkDiscord(discordSubject);
+                }}
+                className="text-sm border border-violet-600 hover:border-violet-700 py-2 px-4 rounded-md text-violet-600 hover:text-violet-700 disabled:border-gray-500 disabled:text-gray-500 hover:disabled:text-gray-500"
+                disabled={!canRemoveAccount}
+              >
+                Unlink Discord
+              </button>
+            ) : (
+              <button
+                className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white"
+                onClick={() => {
+                  linkDiscord();
+                }}
+              >
+                Link Discord
+              </button>
+            )}
+
+            {email ? (
+              <button
+                onClick={() => {
+                  unlinkEmail(email.address);
+                }}
+                className="text-sm border border-violet-600 hover:border-violet-700 py-2 px-4 rounded-md text-violet-600 hover:text-violet-700 disabled:border-gray-500 disabled:text-gray-500 hover:disabled:text-gray-500"
+                disabled={!canRemoveAccount}
+              >
+                Unlink email
+              </button>
+            ) : (
+              <button
+                onClick={linkEmail}
+                className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white"
+              >
+                Connect email
+              </button>
+            )}
+            {wallet ? (
+              <button
+                onClick={() => {
+                  unlinkWallet(wallet.address);
+                }}
+                className="text-sm border border-violet-600 hover:border-violet-700 py-2 px-4 rounded-md text-violet-600 hover:text-violet-700 disabled:border-gray-500 disabled:text-gray-500 hover:disabled:text-gray-500"
+                disabled={!canRemoveAccount}
+              >
+                Unlink wallet
+              </button>
+            ) : (
+              <button
+                onClick={linkWallet}
+                className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white border-none"
+              >
+                Connect wallet
+              </button>
+            )}
+            {phone ? (
+              <button
+                onClick={() => {
+                  unlinkPhone(phone.number);
+                }}
+                className="text-sm border border-violet-600 hover:border-violet-700 py-2 px-4 rounded-md text-violet-600 hover:text-violet-700 disabled:border-gray-500 disabled:text-gray-500 hover:disabled:text-gray-500"
+                disabled={!canRemoveAccount}
+              >
+                Unlink phone
+              </button>
+            ) : (
+              <button
+                onClick={linkPhone}
+                className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white border-none"
+              >
+                Connect phone
+              </button>
+            )}
+
+            <button
+              onClick={() => verifyToken().then(setVerifyResult)}
+              className="text-sm bg-violet-600 hover:bg-violet-700 py-2 px-4 rounded-md text-white border-none"
+            >
+              Verify token on server
+            </button>
+
+            {Boolean(verifyResult) && (
+              <details className="w-full">
+                <summary className="mt-6 font-bold uppercase text-sm text-gray-600">
+                  Server verify result
+                </summary>
+                <pre className="max-w-4xl bg-slate-700 text-slate-50 font-mono p-4 text-xs sm:text-sm rounded-md mt-2">
+                  {JSON.stringify(verifyResult, null, 2)}
+                </pre>
+              </details>
+            )}
+          </div>
+          <div className="space-y-6 max-w-4xl mt-6">
+            <h2 className="text-xl font-bold">Your Wallet</h2>
+            <WalletList />
+          </div>
+          <p className="mt-6 font-bold uppercase text-sm text-gray-600">
+            User object
+          </p>
+          <pre className="max-w-4xl bg-slate-700 text-slate-50 font-mono p-4 text-xs sm:text-sm rounded-md mt-2">
+            {JSON.stringify(user, null, 2)}
+          </pre>
+        </>
+      ) : null}
+    </main>
+  );
+}
