+++ components/Providers.tsx
@@ -1,0 +1,18 @@
+"use client";
+
+import { PrivyProvider } from "@privy-io/react-auth";
+
+export default function Providers({ children }) {
+  return (
+    <PrivyProvider
+      appId={process.env.NEXT_PUBLIC_PRIVY_APP_ID || ""}
+      config={{
+        embeddedWallets: {
+          createOnLogin: "all-users",
+        },
+      }}
+    >
+      {children}
+    </PrivyProvider>
+  );
+}
