+++ package.json
@@ -14,8 +14,8 @@
   "dependencies": {
     "@headlessui/react": "^1.7.3",
     "@heroicons/react": "^2.0.12",
-    "@privy-io/react-auth": "^2.12.0",
-    "@privy-io/server-auth": "1.21.2",
+    "@privy-io/react-auth": "^2.18.1",
+    "@privy-io/server-auth": "^1.28.4",
     "@tailwindcss/forms": "^0.5.3",
     "axios": "^1.9.0",
     "next": "latest",
