+++ tailwind.config.js
@@ -1,20 +1,20 @@
-const defaultTheme = require('tailwindcss/defaultTheme');
+const defaultTheme = require("tailwindcss/defaultTheme");
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
+  content: ["./app/**/*.{js,ts,jsx,tsx}", "./components/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {
       fontFamily: {
-        sans: ['Adelle Sans', ...defaultTheme.fontFamily.sans],
+        sans: ["Adelle Sans", ...defaultTheme.fontFamily.sans],
       },
       colors: {
-        'privy-navy': '#160B45',
-        'privy-light-blue': '#EFF1FD',
-        'privy-blueish': '#D4D9FC',
-        'privy-pink': '#FF8271',
+        "privy-navy": "#160B45",
+        "privy-light-blue": "#EFF1FD",
+        "privy-blueish": "#D4D9FC",
+        "privy-pink": "#FF8271",
       },
     },
   },
-  plugins: [require('@tailwindcss/forms')],
+  plugins: [require("@tailwindcss/forms")],
 };
